refactor: split work into small managable components
