--- src/component/work/projects.css
@@ -1,153 +1,0 @@
-/* projects */
-.projects {
-  display: grid;
-  grid-gap: 0.7rem;
-  grid-template-columns: 1fr;
-}
-.projects .item {
-  position: relative;
-  border: 1px solid #d1d0cd;
-}
-.projects .item img {
-  width: 100%;
-}
-
-@media screen and (min-width: 768px) {
-  .projects {
-    grid-template-columns: repeat(1, 1fr);
-  }
-  .overlay .content p {
-    display: block;
-  }
-}
-
-@media screen and (min-width: 769px) and (max-width: 2499px) {
-  .projects {
-    grid-template-columns: repeat(2, 1fr);
-  }
-}
-
-@media screen and (min-width: 2500px) {
-  .projects {
-    grid-template-columns: repeat(3, 1fr);
-  }
-}
-
-/* Project bars */
-.bar {
-  display: block;
-  position: relative;
-  height: 24px;
-  background-color: #ddd;
-  width: 100%;
-  color: #6a6a6a;
-  z-index: 10;
-  box-shadow: 0 2px 2px 0 rgba(0, 0, 0, 0.14),
-    0 3px 1px -2px rgba(0, 0, 0, 0.12), 0 1px 5px 0 rgba(0, 0, 0, 0.2);
-}
-.bar .fa-blog,
-.bar .fa-firefox {
-  position: absolute;
-  z-index: 12;
-  top: 1px;
-  left: 10px;
-  height: 23px;
-  width: 16px;
-  display: inline-block;
-  color: #fd7e14;
-}
-.bar .window-buttons {
-  background-color: #ff5f57;
-  top: 6px;
-  right: 38px;
-}
-.bar .window-buttons,
-.bar .window-buttons:before {
-  height: 12px;
-  width: 12px;
-  display: inline-block;
-  border-radius: 50%;
-  position: absolute;
-}
-.bar .window-buttons:before {
-  content: " ";
-  left: 16px;
-  background-color: #ffbd2e;
-  top: 0;
-}
-.bar .window-buttons:after {
-  content: " ";
-  left: 32px;
-  background-color: #28ca41;
-  top: 0;
-  height: 12px;
-  width: 12px;
-  display: inline-block;
-  border-radius: 50%;
-  position: absolute;
-}
-.bar .title {
-  display: block;
-  width: 100%;
-  text-align: center;
-  padding: 0.1rem 1rem;
-  font-size: 0.8rem;
-  font-weight: bolder;
-  letter-spacing: -0.3px;
-}
-
-.overlay .content {
-  margin-top: -0.5rem;
-  padding-top: 0.5rem;
-}
-.overlay .content p {
-  font-size: 0.75rem;
-  color: #c6c7c7;
-}
-
-.badges {
-  padding: 0.5rem;
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  justify-content: center;
-  align-content: center;
-}
-
-.item .badge-tech img {
-  width: 25px;
-  height: 25px;
-  padding: 0.2rem;
-  margin-right: 0.25rem;
-}
-
-@media screen and (min-width: 768px) {
-  .overlay .content p {
-    display: block;
-  }
-}
-.overlay {
-  background-color: #404040;
-}
-.overlay .arrow-btn {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  font-size: 0.8rem;
-  text-transform: uppercase;
-}
-.overlay .arrow-btn .arrow-btn-text {
-  background-color: rgba(0, 0, 0, 0.4);
-  padding: 0.5rem;
-}
-.overlay .links {
-  padding: 1rem;
-  display: flex;
-  justify-content: center;
-}
-.overlay svg {
-  padding: 0.05rem 0.35rem;
-  transition: 0.5s ease;
-  font-size: 2rem;
-  height: 35px;
-}
