+++ src/component/work/projects/index.js
@@ -1,0 +1,18 @@
+import React from "react";
+import ProjectCard from "./project";
+import "./projects.css";
+function Projects({ filteredProjects }) {
+  const ProjectCards = () => {
+    return filteredProjects.map((project) => {
+      return <ProjectCard project={project} />;
+    });
+  };
+
+  return (
+    <div className="projects">
+      <ProjectCards />
+    </div>
+  );
+}
+
+export default Projects;
