+++ src/component/work/projects/project/project-bar/bar.css
@@ -1,0 +1,61 @@
+.bar {
+  display: block;
+  position: relative;
+  height: 24px;
+  background-color: #ddd;
+  width: 100%;
+  color: #6a6a6a;
+  z-index: 10;
+  box-shadow: 0 2px 2px 0 rgba(0, 0, 0, 0.14),
+    0 3px 1px -2px rgba(0, 0, 0, 0.12), 0 1px 5px 0 rgba(0, 0, 0, 0.2);
+}
+.bar .fa-blog,
+.bar .fa-firefox {
+  position: absolute;
+  z-index: 12;
+  top: 1px;
+  left: 10px;
+  height: 23px;
+  width: 16px;
+  display: inline-block;
+  color: #fd7e14;
+}
+.bar .window-buttons {
+  background-color: #ff5f57;
+  top: 6px;
+  right: 38px;
+}
+.bar .window-buttons,
+.bar .window-buttons:before {
+  height: 12px;
+  width: 12px;
+  display: inline-block;
+  border-radius: 50%;
+  position: absolute;
+}
+.bar .window-buttons:before {
+  content: " ";
+  left: 16px;
+  background-color: #ffbd2e;
+  top: 0;
+}
+.bar .window-buttons:after {
+  content: " ";
+  left: 32px;
+  background-color: #28ca41;
+  top: 0;
+  height: 12px;
+  width: 12px;
+  display: inline-block;
+  border-radius: 50%;
+  position: absolute;
+}
+.bar .title {
+  display: block;
+  width: 100%;
+  text-align: center;
+  padding: 0.1rem 1rem;
+  font-size: 0.8rem;
+  font-weight: bolder;
+  letter-spacing: -0.3px;
+}
