+++ src/component/work/projects/projects.css
@@ -1,0 +1,38 @@
+.projects {
+  display: grid;
+  grid-gap: 0.7rem;
+  grid-template-columns: 1fr;
+}
+.projects__item {
+  position: relative;
+  border: 1px solid #d1d0cd;
+}
+
+.projects__item-img {
+  width: 100%;
+}
+
+.badge-tech img {
+  width: 25px;
+  height: 25px;
+  padding: 0.2rem;
+  margin-right: 0.25rem;
+}
+
+@media screen and (min-width: 768px) {
+  .projects {
+    grid-template-columns: repeat(1, 1fr);
+  }
+}
+
+@media screen and (min-width: 769px) and (max-width: 2499px) {
+  .projects {
+    grid-template-columns: repeat(2, 1fr);
+  }
+}
+
+@media screen and (min-width: 2500px) {
+  .projects {
+    grid-template-columns: repeat(3, 1fr);
+  }
+}
